+++ origin/tests/TestCase/Controller/Component/ComponentTest.php
@@ -18,7 +18,8 @@ use Origin\Controller\Component\Component;
 use Origin\Controller\Controller;
 use Origin\Controller\Request;
 use Origin\Controller\Response;
-use Origin\Controller\Logger;
+use Origin\Core\Logger;
+use Origin\Controller\Component\SessionComponent;
 
 class MockComponent extends Component
 {
@@ -36,6 +37,13 @@ class ComponentTest extends \PHPUnit\Framework\TestCase
         $this->MockComponent = new MockComponent($Controller);
     }
 
+    public function testGet()
+    {
+        $this->assertNull($this->MockComponent->Session);
+        $this->MockComponent->loadComponent('Session');
+        $this->assertInstanceOf(SessionComponent::class, $this->MockComponent->Session);
+    }
+
     public function testController()
     {
         $this->assertInstanceOf(Controller::class, $this->MockComponent->controller());
@@ -49,5 +57,9 @@ class ComponentTest extends \PHPUnit\Framework\TestCase
     public function testResponse()
     {
         $this->assertInstanceOf(Response::class, $this->MockComponent->response());
+    }
+    public function testLogger()
+    {
+        $this->assertInstanceOf(Logger::class, $this->MockComponent->logger());
     }
 }