+++ origin/tests/TestCase/Controller/RequestTest.php
@@ -108,4 +108,96 @@ class RequestTest extends \PHPUnit\Framework\TestCase
         $request->setInput('{"title":"CNBC","url":""https://www.cnbc.com"}'); // Badd data
         $this->assertEquals([], $request->data());
     }
+
+    public function testQuery()
+    {
+        $request = new MockRequest();
+        $request->query('key', 'value');
+        $this->assertEquals('value', $request->query('key'));
+        $this->assertEquals(['key'=>'value'], $request->query());
+        $this->assertNull($request->query('fozzy'));
+    }
+
+    public function testData()
+    {
+        $request = new MockRequest();
+        $request->data('key', 'value');
+        $this->assertEquals('value', $request->data('key'));
+        $this->assertEquals(['key'=>'value'], $request->data());
+        $this->assertNull($request->data('fozzy'));
+    }
+
+    public function testParams()
+    {
+        $request = new MockRequest();
+        $request->params('key', 'value');
+        $this->assertEquals('value', $request->params('key'));
+        $this->assertNotEmpty($request->params());
+        $this->assertNull($request->params('fozzy'));
+    }
+
+    public function testMethod()
+    {
+        $request = new MockRequest();
+        $_SERVER['REQUEST_METHOD'] = 'POST';
+        $this->assertEquals('POST', $request->method());
+        $_SERVER['REQUEST_METHOD'] = '';
+    }
+
+    public function testFiles()
+    {
+        $_FILES = ['file'=>'dummy file'];
+        $request = new MockRequest();
+        $this->assertEquals('dummy file', $request->data('file'));
+    }
+
+    public function testCookie()
+    {
+        $expected = $_COOKIE = ['foo'=>'PLgTmbAZY5BZjA9tQbp5h50GI1wbXZldGV9cERAx5c6C4nvWBH8Ouc+tGbX+1mfv'];
+   
+        $request = new MockRequest();
+        $this->assertEquals($expected, $request->cookie());
+        $this->assertEquals('This is a test', $request->cookie('foo'));
+    }
+    public function testHeaders()
+    {
+        $request = new MockRequest();
+        $request->header('WWW-Authenticate', 'Negotiate');
+        
+        $this->assertEquals('Negotiate', $request->header('WWW-Authenticate'));
+        $this->assertEquals('Negotiate', $request->header('www-Authenticate')); // PSR friendly
+        $this->assertEquals(['WWW-Authenticate'=>'Negotiate'], $request->headers());
+
+        $this->assertEquals('', $request->header('secret'));
+    }
+
+    /**
+     * @depends testHeaders
+     */
+    public function testAcceptLanguage()
+    {
+        $request = new MockRequest();
+        $request->header('Accept-Language', 'en-GB,en;q=0.9,es;q=0.8');
+        $this->assertTrue($request->acceptLanguage('en'));
+        $this->assertEquals(['en_GB','en','es'], $request->acceptLanguage());
+    }
+    /**
+       * @depends testHeaders
+       */
+    public function testAccepts()
+    {
+        $request = new MockRequest();
+        $this->assertFalse($request->accepts('application/json'));
+        $request = new MockRequest('/controller/action.json');
+        $this->assertTrue($request->accepts('application/json'));
+        $request = new MockRequest();
+        $request->params('json', true);
+        $this->assertTrue($request->accepts('application/json'));
+
+        $request = new MockRequest();
+        $request->header('Accept', 'text/html,application/xhtml+xml,application/xml;q=0.9,image/webp,image/apng,*/*;q=0.8,application/signed-exchange;v=b3');
+        $accepts = $request->accepts();
+        $this->assertEquals('text/html', $accepts[0]);
+        $this->assertTrue($request->accepts(['application/xml','application/json']));
+    }
 }