+++ origin/tests/TestCase/Model/DatasourceTest.php
@@ -20,6 +20,7 @@ use Origin\Model\Driver\MySQLDriver;
 
 use PDOException;
 use Origin\Model\Exception\DatasourceException;
+use Origin\Exception\Exception;
 
 class DatasourceTest extends \PHPUnit\Framework\TestCase
 {
@@ -255,7 +256,79 @@ class DatasourceTest extends \PHPUnit\Framework\TestCase
         $this->assertTrue($result);
     }
 
+    public function testTables()
+    {
+        $tables = $this->connection->tables();
+        $this->assertTrue(in_array('authors', $tables));
+        $this->assertTrue(in_array('articles', $tables));
+    }
 
+    public function testCreateTable()
+    {
+        $this->connection->execute('DROP TABLE IF EXISTS foo;');
+
+        $schema = [
+            'id' => ['type'=>'integer','autoIncrement'=>true,'key'=>'primary'],
+            'name' => ['type'=>'string','default'=>'placeholder'],
+            'description' => ['type'=>'text','null'=>false],
+            'age' => ['type'=>'integer','default'=>1234],
+            'bi' => ['type'=>'biginteger'],
+            'fn' => ['type'=>'float','precision'=>2],
+            'dn' => ['type'=>'decimal','precision'=>2],
+            'dt' => ['type'=>'datetime'],
+            'ts' => ['type'=>'timestamp'],
+            't' => ['type'=>'time'],
+            'd' => ['type'=>'date'],
+            'bf' => ['type'=>'binary'],
+            'bool' => ['type'=>'boolean'],
+        ];
+        $result = $this->connection->createTable('foo', $schema);
+        $this->assertTrue($this->connection->execute($result));
+
+        $schema = [
+            'bookmarks_id' => ['type'=>'integer','key'=>'primary'],
+            'tags_id' => ['type'=>'integer','key'=>'primary'],
+        ];
+        $result = $this->connection->createTable('bar', $schema);
+        $this->assertContains('bookmarks_id INT,', $result);
+        $this->assertContains('tags_id INT,', $result);
+        $this->assertContains('PRIMARY KEY (bookmarks_id,tags_id)', $result);
+
+        $this->expectException(Exception::class);
+        $schema = [
+            'key' => ['type'=>'object']
+        ];
+        $result = $this->connection->createTable('foo/bar', $schema);
+    }
+
+    /**
+     * @depends testCreateTable
+     *
+     * @return void
+     */
+    public function testSchema()
+    {
+        $schema = $this->connection->schema('foo');
+        $this->assertEquals('integer', $schema['id']['type']);
+        $this->assertEquals('primary', $schema['id']['key']);
+        $this->assertTrue($schema['id']['autoIncrement']);
+        $this->assertEquals('string', $schema['name']['type']);
+        $this->assertEquals('placeholder', $schema['name']['default']);
+        $this->assertEquals('text', $schema['description']['type']);
+        $this->assertEquals('integer', $schema['age']['type']);
+        $this->assertEquals(1234, $schema['age']['default']);
+        $this->assertEquals('biginteger', $schema['bi']['type']);
+        $this->assertEquals('float', $schema['fn']['type']);
+        $this->assertEquals(2, $schema['fn']['precision']);
+        $this->assertEquals('decimal', $schema['dn']['type']);
+        $this->assertEquals(2, $schema['dn']['precision']);
+        $this->assertEquals('datetime', $schema['dt']['type']);
+        $this->assertEquals('timestamp', $schema['ts']['type']);
+        $this->assertEquals('date', $schema['d']['type']);
+        $this->assertEquals('time', $schema['t']['type']);
+        $this->assertEquals('binary', $schema['bf']['type']);
+        $this->assertEquals('boolean', $schema['bool']['type']);
+    }
 
     public function createTables()
     {
@@ -266,8 +339,6 @@ class DatasourceTest extends \PHPUnit\Framework\TestCase
             $this->connection->execute($statement);
         }
     }
-
-
 
     private function getStatements()
     {